requirements.txt
music21
midi2audio
torchvision
lightning
torchinfo
librosa
scikit-learn
fire
pyMV2H
swalign==0.3.6
wandb
rich